+++ Cesium3DTilesSelection/src/Tileset.cpp
@@ -7,6 +7,7 @@
 #include <Cesium3DTilesSelection/TileID.h>
 #include <Cesium3DTilesSelection/TileOcclusionRendererProxy.h>
 #include <Cesium3DTilesSelection/Tileset.h>
+#include <Cesium3DTilesSelection/TilesetMetadata.h>
 #include <Cesium3DTilesSelection/spdlog-cesium.h>
 #include <CesiumAsync/AsyncSystem.h>
 #include <CesiumGeospatial/Cartographic.h>
@@ -86,15 +87,19 @@ Tileset::Tileset(
           ionAccessToken,
           ionAssetEndpointUrl)} {}
 
-CesiumAsync::SharedFuture<void>& Tileset::getAsyncDestructionCompleteEvent() {
-  return this->_pTilesetContentManager->getAsyncDestructionCompleteEvent();
-}
-
 Tileset::~Tileset() noexcept {
   this->_pTilesetContentManager->unloadAll();
   if (this->_externals.pTileOcclusionProxyPool) {
     this->_externals.pTileOcclusionProxyPool->destroyPool();
   }
+}
+
+CesiumAsync::SharedFuture<void>& Tileset::getAsyncDestructionCompleteEvent() {
+  return this->_pTilesetContentManager->getAsyncDestructionCompleteEvent();
+}
+
+CesiumAsync::SharedFuture<void>& Tileset::getRootTileAvailableEvent() {
+  return this->_pTilesetContentManager->getRootTileAvailableEvent();
 }
 
 const std::vector<Credit>& Tileset::getTilesetCredits() const noexcept {
@@ -453,6 +458,61 @@ int64_t Tileset::getTotalDataBytes() const noexcept {
   return this->_pTilesetContentManager->getTotalDataUsed();
 }
 
+const TilesetMetadata* Tileset::getMetadata(const Tile* pTile) const {
+  if (pTile == nullptr) {
+    pTile = this->getRootTile();
+  }
+
+  while (pTile != nullptr) {
+    const TileExternalContent* pExternal =
+        pTile->getContent().getExternalContent();
+    if (pExternal)
+      return &pExternal->metadata;
+    pTile = pTile->getParent();
+  }
+
+  return nullptr;
+}
+
+CesiumAsync::Future<const TilesetMetadata*> Tileset::loadMetadata() {
+  return this->getRootTileAvailableEvent().thenInMainThread(
+      [pManager = this->_pTilesetContentManager,
+       pAssetAccessor = this->_externals.pAssetAccessor,
+       asyncSystem =
+           this->getAsyncSystem()]() -> Future<const TilesetMetadata*> {
+        Tile* pRoot = pManager->getRootTile();
+        assert(pRoot);
+
+        TileExternalContent* pExternal =
+            pRoot->getContent().getExternalContent();
+        if (!pExternal) {
+          return asyncSystem.createResolvedFuture<const TilesetMetadata*>(
+              nullptr);
+        }
+
+        TilesetMetadata& metadata = pExternal->metadata;
+        if (!metadata.schemaUri) {
+          // No schema URI, so the metadata is ready to go.
+          return asyncSystem.createResolvedFuture<const TilesetMetadata*>(
+              &metadata);
+        }
+
+        return metadata.loadSchemaUri(asyncSystem, pAssetAccessor)
+            .thenInMainThread(
+                [pManager, pAssetAccessor]() -> const TilesetMetadata* {
+                  Tile* pRoot = pManager->getRootTile();
+                  assert(pRoot);
+
+                  TileExternalContent* pExternal =
+                      pRoot->getContent().getExternalContent();
+                  if (!pExternal) {
+                    return nullptr;
+                  }
+                  return &pExternal->metadata;
+                });
+      });
+}
+
 static void markTileNonRendered(
     TileSelectionState::Result lastResult,
     Tile& tile,
@@ -781,13 +841,16 @@ Tileset::TraversalDetails Tileset::_visitTileIfNeeded(
   this->_frustumCull(tile, frameState, cullWithChildrenBounds, cullResult);
   this->_fogCull(frameState, distances, cullResult);
 
-  if (this->_options.forbidHoles && !cullResult.shouldVisit &&
-      tile.getRefine() == TileRefine::Replace &&
-      tile.getUnconditionallyRefine()) {
+  if (!cullResult.shouldVisit && tile.getUnconditionallyRefine()) {
     // Unconditionally refined tiles must always be visited in forbidHoles
     // mode, because we need to load this tile's descendants before we can
-    // render any of its siblings.
-    cullResult.shouldVisit = true;
+    // render any of its siblings. An unconditionally refined root tile must be
+    // visited as well, otherwise we won't load anything at all.
+    if ((this->_options.forbidHoles &&
+         tile.getRefine() == TileRefine::Replace) ||
+        tile.getParent() == nullptr) {
+      cullResult.shouldVisit = true;
+    }
   }
 
   if (!cullResult.shouldVisit) {