+++ Cesium3DTilesSelection/src/TilesetContentManager.cpp
@@ -37,8 +37,9 @@ struct ContentKindSetter {
     tileContent.setContentKind(content);
   }
 
-  void operator()(TileExternalContent content) {
-    tileContent.setContentKind(content);
+  void operator()(TileExternalContent&& content) {
+    tileContent.setContentKind(
+        std::make_unique<TileExternalContent>(std::move(content)));
   }
 
   void operator()(CesiumGltf::Model&& model) {
@@ -598,7 +599,12 @@ TilesetContentManager::TilesetContentManager(
       _tilesDataUsed{0},
       _destructionCompletePromise{externals.asyncSystem.createPromise<void>()},
       _destructionCompleteFuture{
-          this->_destructionCompletePromise.getFuture().share()} {}
+          this->_destructionCompletePromise.getFuture().share()},
+      _rootTileAvailablePromise{externals.asyncSystem.createPromise<void>()},
+      _rootTileAvailableFuture{
+          this->_rootTileAvailablePromise.getFuture().share()} {
+  this->_rootTileAvailablePromise.resolve();
+}
 
 TilesetContentManager::TilesetContentManager(
     const TilesetExternals& externals,
@@ -622,7 +628,10 @@ TilesetContentManager::TilesetContentManager(
       _tilesDataUsed{0},
       _destructionCompletePromise{externals.asyncSystem.createPromise<void>()},
       _destructionCompleteFuture{
-          this->_destructionCompletePromise.getFuture().share()} {
+          this->_destructionCompletePromise.getFuture().share()},
+      _rootTileAvailablePromise{externals.asyncSystem.createPromise<void>()},
+      _rootTileAvailableFuture{
+          this->_rootTileAvailablePromise.getFuture().share()} {
   if (!url.empty()) {
     this->notifyTileStartLoading(nullptr);
 
@@ -720,13 +729,16 @@ TilesetContentManager::TilesetContentManager(
                   TilesetLoadType::TilesetJson,
                   errorCallback,
                   std::move(result));
+              thiz->_rootTileAvailablePromise.resolve();
             })
         .catchInMainThread([thiz](std::exception&& e) {
           thiz->notifyTileDoneLoading(nullptr);
           SPDLOG_LOGGER_ERROR(
               thiz->_externals.pLogger,
-              "An unexpected error occurs when loading tile: {}",
+              "An unexpected error occurred when loading tile: {}",
               e.what());
+          thiz->_rootTileAvailablePromise.reject(
+              std::runtime_error("Root tile failed to load."));
         });
   }
 }
@@ -755,7 +767,10 @@ TilesetContentManager::TilesetContentManager(
       _tilesDataUsed{0},
       _destructionCompletePromise{externals.asyncSystem.createPromise<void>()},
       _destructionCompleteFuture{
-          this->_destructionCompletePromise.getFuture().share()} {
+          this->_destructionCompletePromise.getFuture().share()},
+      _rootTileAvailablePromise{externals.asyncSystem.createPromise<void>()},
+      _rootTileAvailableFuture{
+          this->_rootTileAvailablePromise.getFuture().share()} {
   if (ionAssetID > 0) {
     auto authorizationChangeListener = [this](
                                            const std::string& header,
@@ -792,13 +807,16 @@ TilesetContentManager::TilesetContentManager(
                   TilesetLoadType::CesiumIon,
                   errorCallback,
                   std::move(result));
+              thiz->_rootTileAvailablePromise.resolve();
             })
         .catchInMainThread([thiz](std::exception&& e) {
           thiz->notifyTileDoneLoading(nullptr);
           SPDLOG_LOGGER_ERROR(
               thiz->_externals.pLogger,
-              "An unexpected error occurs when loading tile: {}",
+              "An unexpected error occurred when loading tile: {}",
               e.what());
+          thiz->_rootTileAvailablePromise.reject(
+              std::runtime_error("Root tile failed to load."));
         });
   }
 }
@@ -806,6 +824,11 @@ TilesetContentManager::TilesetContentManager(
 CesiumAsync::SharedFuture<void>&
 TilesetContentManager::getAsyncDestructionCompleteEvent() {
   return this->_destructionCompleteFuture;
+}
+
+CesiumAsync::SharedFuture<void>&
+TilesetContentManager::getRootTileAvailableEvent() {
+  return this->_rootTileAvailableFuture;
 }
 
 TilesetContentManager::~TilesetContentManager() noexcept {